Move ElGamal class into its own module
